feat: implement PlateauModelUtil class for mesh code extraction and coordinate transformation
