+++ src/index.ts
@@ -7,6 +7,7 @@ export * from "./JapanStandardRegionalMeshUtil.js";
 export * as JapanStandardRegionalMeshUtil from "./JapanStandardRegionalMeshUtil.js";
 export * from "./LatitudeLongitude.js";
 export { PlateauModelLoader } from "./PlateauModelLoader.js";
+export * from "./PlateauModelUtil.js";
 export * from "./PositionUtil.js";
 /**
  * @deprecated This namespace import is deprecated. Instead, import individual functions or
